migrated contract connection
migrated contract connection to an internal style class to allow for internal abstraction in preparation for the Multi Service contract connection
