+++ Connectors/ActiveMQ/ConsumerInstance.cs
@@ -1,9 +1,4 @@
 ﻿using Apache.NMS;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MQContract.ActiveMQ
 {
